+++ hegel-macros/src/lib.rs
@@ -1,10 +1,11 @@
+mod composite;
 mod enum_gen;
 mod hegel_test;
 mod struct_gen;
 mod utils;
 
 use proc_macro::TokenStream;
-use syn::{Data, DeriveInput, parse_macro_input};
+use syn::{Data, DeriveInput, ItemFn, parse_macro_input};
 
 /// Derive a generator for a struct or enum.
 ///
@@ -63,12 +64,12 @@ use syn::{Data, DeriveInput, parse_macro_input};
 /// }
 /// ```
 #[proc_macro_derive(Generator)]
-pub fn derive_generate(input: TokenStream) -> TokenStream {
+pub fn derive_generator(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
 
     match &input.data {
-        Data::Struct(data) => struct_gen::derive_struct_generate(&input, data),
-        Data::Enum(data) => enum_gen::derive_enum_generate(&input, data),
+        Data::Struct(data) => struct_gen::derive_struct_generator(&input, data),
+        Data::Enum(data) => enum_gen::derive_enum_generator(&input, data),
         Data::Union(_) => syn::Error::new_spanned(&input, "Generator cannot be derived for unions")
             .to_compile_error()
             .into(),
@@ -100,4 +101,10 @@ pub fn derive_generate(input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 pub fn test(attr: TokenStream, item: TokenStream) -> TokenStream {
     hegel_test::expand_test(attr.into(), item.into()).into()
+}
+
+#[proc_macro_attribute]
+pub fn composite(_attr: TokenStream, item: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(item as ItemFn);
+    composite::expand_composite(input).into()
 }